% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_map.R
\name{elevation_map}
\alias{elevation_map}
\title{Elevation Map}
\usage{
elevation_map(locations, z = 6, dsn)
}
\arguments{
\item{locations}{A dataframe --> set map boundaries}

\item{z}{The zoom level. Range \link{1,14}}

\item{dsn}{data source name of Polygonal file -> shape/ boundary of map.
Download .zip file from \href{https://www.diva-gis.org/gdata}{diva-gis}.
Corresponding file format = SHAPEFILE (.shp)}
}
\description{
The function creates elevation map with 2 Raster Layers
}
