% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehuynh_theme.R
\name{lehuynh_theme}
\alias{lehuynh_theme}
\title{Le-Huynh's ggplot2 theme}
\usage{
lehuynh_theme(base_size = 11, base_family = "", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
An object as returned by \code{\link[ggplot2:theme]{ggplot2::theme()}}
}
\description{
Le-Huynh's ggplot2 theme: white background, black axis, black text
}
\examples{

library(ggplot2)

fig <- ggplot(mtcars, aes(y = mpg, x = disp)) +
    geom_point(aes(colour = factor(cyl)))

fig

fig + lehuynh_theme()
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}, \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}
}
