% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MinMaxScaling.R
\name{MinMaxScaling}
\alias{MinMaxScaling}
\title{Min-Max Standardization}
\usage{
MinMaxScaling(x, y = x)
}
\arguments{
\item{x}{numeric object}

\item{y}{numeric object, default: y = x}
}
\value{
numeric object
}
\description{
Normalize / Standardize / Scale the data to the fixed range from 0 to 1.
The minimum value of data gets transformed into 0.
The maximum value gets transformed into 1.
Other values get transformed into decimals between 0 and 1.
}
\examples{
dat1 = seq(from = 5, to = 30, lenght.out = 6)

MinMaxScaling(dat1)

dat2 = c(7, 13, 22)
MinMaxScaling(x = dat2, y = dat1)
}
