% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ngrams_filter.R
\name{ngrams_filter}
\alias{ngrams_filter}
\title{Filter and generate N-Grams from text data}
\usage{
ngrams_filter(
  data,
  group_column,
  group_name,
  text_column,
  ngrams,
  user_defined_stopwords = NULL
)
}
\arguments{
\item{data}{A data frame containing the dataset to be processed.}

\item{group_column}{A character string specifying the name of the column
used to filter the data.}

\item{group_name}{A character string specifying the value within
the group column to filter the data by.}

\item{text_column}{A character string specifying the name of the column
containing text data to be tokenized into n-grams.}

\item{ngrams}{An integer specifying the number of words in the n-grams
to be generated.}

\item{user_defined_stopwords}{A character vector of additional stopwords
to be removed from the n-grams. Default is NULL.}
}
\value{
A data frame with the filtered data and generated n-grams,
excluding the specified stopwords.
}
\description{
Filter a dataset based on a specified column and group value,
generate n-grams from a specified text column,
then remove standard and user-defined stopwords from the n-grams.
}
\examples{

library(janeaustenr)

austen_books() \%>\%
          ngrams_filter(group_column = "book",
                        group_name = "Pride & Prejudice",
                        text_column = "text",
                        ngrams = 2)

austen_books() \%>\%
          ngrams_filter(group_column = "book",
                        group_name = "Pride & Prejudice",
                        text_column = "text",
                        ngrams = 2,
                        user_defined_stopwords = c("chapter", 1:50))
}
