% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_networkD3.R
\name{plot_networkD3}
\alias{plot_networkD3}
\title{Plot network using NetworkD3}
\usage{
plot_networkD3(
  data,
  col1 = "word1",
  col2 = "word2",
  plot_type = c("whole_network", "biggest_component_one_color",
    "biggest_component_community_color"),
  threshold,
  node_size = 20,
  edges_width = 10,
  opacity = 1,
  font_size = 15,
  ...
)
}
\arguments{
\item{data}{A data frame containing the edge list for the network.}

\item{col1, col2}{The name of two columns containing the symbolic edge list.
Default is "word1" and "word2", respectively.}

\item{plot_type}{A character string specifying the type of plot to generate.
Options are "whole_network", "biggest_component_one_color",
and "biggest_component_community_color". Default is "whole_network".}

\item{threshold}{An integer specifying the minimum frequency of edges
to be included in the network.}

\item{node_size}{A numeric value specifying the base size of the nodes.
Default is 20.}

\item{edges_width}{A numeric value specifying the base width of the edges.
Default is 10.}

\item{opacity}{A numeric value specifying the opacity of the graph elements.
Default is 1.}

\item{font_size}{A numeric value specifying the font size of the node labels.
Default is 15.}

\item{...}{Additional arguments passed to \code{\link[networkD3:forceNetwork]{networkD3::forceNetwork()}}.}
}
\value{
An interactive network plot created using \code{networkD3}.
}
\description{
This function processes the input data to create an igraph object and
then generates an interactive network plot based on the specified plot type.
The plot can show the entire network, the largest component with a single
color, or the largest component with different colors based on
community detection.
Node size and edge width are scaled based on node degree and edge weight,
respectively.
}
\examples{
library(janeaustenr)

data <- austen_books() \%>\%
          ngrams_filter(group_column = "book",
                        group_name = "Pride & Prejudice",
                        text_column = "text",
                        ngrams = 2)

# The whole network plot
plot_networkD3(data = data,
               threshold = 10)

# The biggest component plot with one color
plot_networkD3(data = data,
               plot_type = "biggest_component_one_color",
               threshold = 10)

# The biggest component plot with community based color
plot_networkD3(data = data,
               plot_type = "biggest_component_community_color",
               threshold = 10)

}
