% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_excel.R
\name{import_excel}
\alias{import_excel}
\title{Import excel file with multiple sheets}
\usage{
import_excel(file_path, output = c("list", "dataframe"))
}
\arguments{
\item{file_path}{path to the xls/xlsx file.}

\item{output}{type of output, \code{list} or \code{data-frame}.}
}
\value{
\itemize{
\item \code{output = "list"}, a \emph{named} \code{list} of imported sheets
\item \code{output = "dataframe"}, multiple data frames in the .GlobalEnv
}
}
\description{
This function imports an excel file with multiple sheets and returns
a \emph{named} \code{list} of imported sheets or multiple data frames in the
.GlobalEnv. The data frames are named after the sheet names.
}
\examples{
\dontrun{
path <- here::here("data/data.xlsx")

mydata <- import_excel(file_path = path,
                      output = "list")

import_excel(file_path = path,
             output = "dataframe")
}
}
