% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lehuynh_theme.R
\name{lehuynh_theme}
\alias{lehuynh_theme}
\title{Le's ggplot2 theme}
\usage{
lehuynh_theme(base_size = 11, base_family = "", ...)
}
\arguments{
\item{base_size}{Base font size}

\item{base_family}{Base font family}

\item{...}{Passed to \code{\link[ggplot2:theme]{ggplot2::theme()}}}
}
\value{
An object as returned by \code{\link[ggplot2:theme]{ggplot2::theme()}}
}
\description{
Le's ggplot2 theme: white background, black axis, black text
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
library(ggplot2)
mtcars$cyl <- factor(mtcars$cyl)
ggplot(mtcars, aes(y = mpg, x = disp, color = cyl)) +
    geom_point() + lehuynh_theme()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[ggplot2:theme]{ggplot2::theme()}}, \code{\link[ggplot2:ggtheme]{ggplot2::theme_bw()}}
}
