% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppc_brms.R
\name{ppc_brms}
\alias{ppc_brms}
\title{Fitted versus observed plot for brmsfit Objects}
\usage{
ppc_brms(
  object,
  xtitle = "Observed value",
  ytitle = "Fitted value",
  dy = c(0.1, 0.1),
  dx = c(0.1, 0.1),
  ...
)
}
\arguments{
\item{object}{An object of class \strong{brmsfit}}

\item{xtitle}{The text for the x-axis title}

\item{ytitle}{The text for the y-axis title}

\item{dy}{Distance from plot to y-axis}

\item{dx}{Distance from plot to x-axis}

\item{...}{Passed to \code{\link[=lehuynh_theme]{lehuynh_theme()}}}
}
\value{
A plot
}
\description{
Plot fitted versus observed values, including confidence interval
(gray area) around best fit line (linear regression line) and
prediction interval (dashed line).
}
\examples{
\dontrun{
library(ggplot2)
library(brms)

mod <- brm(count ~ zAge + zBase * Trt + (1|patient) + (1|obs),
           data = epilepsy,
           family = poisson())

ppc_brms(mod)
ppc_brms(mod, dy = c(0.02, 0.1), dx = c(0.005, 0.1))
}
}
