% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytuesday.R
\name{tidytuesday}
\alias{tidytuesday}
\title{Create a new folder for \emph{#tidytuesday} challenge}
\usage{
tidytuesday(dir_name)
}
\arguments{
\item{dir_name}{name of the folder, a character string}
}
\value{

}
\description{
\emph{\href{https://github.com/rfordatascience/tidytuesday}{#tidytuesday}} challenge
}
\details{
The folder includes:
\itemize{
\item \strong{README.md}: plots for \emph{#tidytuesday} challenge
\item different folders to hold all \emph{data}, \emph{code}, \emph{plots} of data analysis
\item templates: \strong{code.R}
}
}
