% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidytuesday.R
\name{tidytuesday}
\alias{tidytuesday}
\title{Create a new folder for \emph{#tidytuesday} challenge}
\usage{
tidytuesday(year, week)
}
\arguments{
\item{year}{numeric entry for year}

\item{week}{numeric entry for the week of interest, from 1 to 52}
}
\value{
A folder containing folders and files for \emph{#tidytuesday} challenge.
}
\description{
\emph{\href{https://github.com/rfordatascience/tidytuesday}{#tidytuesday}} challenge
}
\details{
The folder includes:
\itemize{
\item \strong{README.md}: plots for \emph{#tidytuesday} challenge
\item different folders to hold all \emph{data}, \emph{code}, \emph{plots} of data analysis
\item templates
}
}
\examples{
\dontshow{if (FALSE) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
tidytuesday(2023, 42)
\dontshow{\}) # examplesIf}
}
